@@ -6,6 +6,7 @@
     "@testing-library/jest-dom": "^5.16.1",
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
+    "@types/file-saver": "^2.0.5",
     "@types/jest": "^27.4.1",
     "@types/node": "^17.0.21",
     "@types/react": "^17.0.39",
@@ -13,17 +14,20 @@
     "@types/react-router-bootstrap": "^0.24.5",
     "@types/uuid": "^8.3.4",
     "bootstrap": "^5.1.3",
+    "draft-js": "^0.11.7",
+    "file-saver": "^2.0.5",
     "final-form": "^4.20.6",
     "firebase": "^9.6.7",
+    "firebase-functions": "^3.20.0",
     "react": "^17.0.2",
     "react-bootstrap": "^2.1.1",
     "react-dom": "^17.0.2",
+    "react-dropzone": "^12.0.5",
     "react-final-form": "^6.5.8",
     "react-router-bootstrap": "^0.25.0",
     "react-router-dom": "^5.3.0",
     "react-scripts": "5.0.0",
     "store": "^2.0.12",
-    "typescript": "^4.5.5",
     "uuid": "^8.3.2",
     "web-vitals": "^2.1.3"
   },
@@ -46,5 +50,11 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@types/draft-js": "^0.11.9",
+    "@typescript-eslint/eslint-plugin": "^5.17.0",
+    "@typescript-eslint/parser": "^5.17.0",
+    "typescript": "^4.6.3"
   }
 }
