@@ -1,22 +1,22 @@
-import { DeckAppType, ItemType } from "../types/types";
-
-const store = require('store');
+import { ItemType } from "../types/types";
 
 const useScheduleAlgorithm = () => {
 
-  const calculateInterval = (date: Date, timestamp: string | Date) => {
+  const calculateInterval = (date: Date, timestamp: string) => {
     return Math.floor((date.getTime() - new Date(timestamp).getTime()) / 60000);
   };
 
-  const selectRepetitionCards = (deckItems: ItemType[]) => {
+  const getCardsToLearning = (deckItems: ItemType[]) => {
     const currentDate = new Date();
-
     let sortedDeck: ItemType[] = [];
 
     deckItems.forEach(item => {
       switch (item.rating) {
+      case 1:
+        sortedDeck = [...sortedDeck, item];
+        break;
       case 2:
-        if (calculateInterval(currentDate, item.date) >= 1) {
+        if (calculateInterval(currentDate, item.date) >= 2) {
           sortedDeck = [...sortedDeck, item];
         }
         break;
@@ -25,46 +25,30 @@ const useScheduleAlgorithm = () => {
           sortedDeck = [...sortedDeck, item];
         }
         break;
-      case 4:
-        if (calculateInterval(currentDate, item.date) >= 1440) {
+      case 33:
+        if (calculateInterval(currentDate, item.date) >= 60) {
           sortedDeck = [...sortedDeck, item];
         }
         break;
-      case 5:
-        if (calculateInterval(currentDate, item.date) >= 10080) {
+      case 4:
+        if (calculateInterval(currentDate, item.date) >= 300) {
+          sortedDeck = [...sortedDeck, item];
+        }
+        break;
+      case 44:
+        if (calculateInterval(currentDate, item.date) >= 7200) {
           sortedDeck = [...sortedDeck, item];
         }
         break;
       default:
-        sortedDeck = [...deckItems];
+        return
       }
     });
 
     return sortedDeck;
-    
   };
 
-  const updateRepetitionInStorage = (deckItems: ItemType[], deckTitle: string) => {
-    let deck = store.get(deckTitle);
-    
-    for (let i = 0; i < deckItems.length; i++) {
-
-      for (let j = 0; j < deck.items.length; j++) {
-        
-        if (deck.items[j].id === deckItems[i].id) {
-          
-          deck.items[j].rating = deckItems[i].rating;
-          deck.items[j].date = deckItems[i].date;
-
-        }
-      }
-    }
-    
-    store.set(deckTitle, deck);
-  
-  };
-
-  return { selectRepetitionCards, updateRepetitionInStorage };
+  return {  getCardsToLearning };
 };
 
 export default useScheduleAlgorithm;
