@@ -1,12 +1,15 @@
 import { FC, FormEvent, useRef, useState } from 'react';
 import { Form, Button, Card, Alert } from 'react-bootstrap';
 import { Link, useHistory } from 'react-router-dom';
+import { collection } from 'firebase/firestore';
+import { db } from '../../firebase/firebase-config';
 import { useAuth } from '../contexts/auth-context';
 
 const Signup: FC = () => {
   const [error, setError] = useState<string>('');
   const [loading, setLoading] = useState<boolean>(false);
 
+  const nicknameRef = useRef<HTMLInputElement>(null);
   const emailRef = useRef<HTMLInputElement>(null);
   const passwordRef = useRef<HTMLInputElement>(null);
   const passwordConfirmRef = useRef<HTMLInputElement>(null);
@@ -23,7 +26,7 @@ const Signup: FC = () => {
     try {
       setError('');
       setLoading(true);
-      await signup(emailRef.current?.value, passwordRef.current?.value);
+      await signup(nicknameRef.current?.value, emailRef.current?.value, passwordRef.current?.value);
       history.push("/");
     } catch {
       setError("Failed to create an account");
@@ -38,6 +41,10 @@ const Signup: FC = () => {
           <h2 className='text-center mb-4'>Sign up</h2>
           {error && <Alert variant="danger">{error}</Alert>}
           <Form onSubmit={handleSubmit}>
+            <Form.Group id="nickname">
+              <Form.Label>Nickname</Form.Label>
+              <Form.Control className="mb-2" type="text" ref={nicknameRef} required/>
+            </Form.Group>
             <Form.Group id="email">
               <Form.Label>Email</Form.Label>
               <Form.Control className="mb-2" type="email" ref={emailRef} required/>
@@ -50,7 +57,7 @@ const Signup: FC = () => {
               <Form.Label>Password Confirmation</Form.Label>
               <Form.Control type="password" ref={passwordConfirmRef} required/>
             </Form.Group>
-            <Button disabled={loading} className="w-100 mt-4" type="submit">Sign Up</Button>
+            <Button disabled={loading} className="w-100 mt-3" type="submit">Sign Up</Button>
           </Form>
         </Card.Body>
       </Card>
