@@ -3,50 +3,53 @@ import { Button } from "react-bootstrap";
 import { LinkContainer } from "react-router-bootstrap";
 import { useAuth } from "../../../contexts/auth-context";
 import useFirestore from "../../../../hooks/firestore";
-import useScheduleAlgorithm from "../../../../hooks/schedule-algorithm";
-import { DeckAppType } from "../../../../types/types";
+import { DeckAppType, ItemType } from "../../../../types/types";
 
 type DeckListItemPropsType = {
   deck: DeckAppType
   onDeleteDeck: (deckId: string) => void
+  exportDeck: (deck: Array<ItemType>, deckTitle: string) => void
 }
 
 const DeckListItem: FC<DeckListItemPropsType> = ({
   deck,
-  onDeleteDeck
+  onDeleteDeck,
+  exportDeck
 }) => {
 
-  const { setEditMode, setCurrentDeck } = useAuth();
+  const { setEditMode, setCurrentDeckId } = useAuth();
   const { getDeckFromFirestore} = useFirestore();
-  const { selectRepetitionCards } = useScheduleAlgorithm();
-
-  const onSelectDeck = (deckId: string, editMode: boolean) => {
-    setEditMode(editMode);
-    if (!editMode) {
-      getDeckFromFirestore(deckId)
-        .then(response => setCurrentDeck(response))
-        .catch((e) => console.log(e))
-    }
-  };
 
   return (
     <li className="list-group-item d-flex justify-content-between">
       <LinkContainer 
         to="/main-table-page" 
-        onClick={ () => setCurrentDeck(deck)}
-        style={ { 'cursor': 'pointer', 'marginTop': '15px' } }
+        onClick={ () => setCurrentDeckId(deck.deckId) }
+        style={ { 'cursor': 'pointer' } }
       >
-        <span className="list-group-item-label" data-deck={ deck.deckTitle }>
+        <div 
+          className="d-flex align-items-center" 
+          data-deck={ deck.deckTitle }
+        >
           { deck.deckTitle }
-        </span>
+        </div>
       </LinkContainer>
       <div className="button-container">
-        <LinkContainer to="/edit-deck-page" onClick={ () => setCurrentDeck(deck)}>
-          <Button variant="secondary">Edit</Button>
+        <LinkContainer
+          to="/main-table-page" 
+          onClick={ () => setCurrentDeckId(deck.deckId) }
+        >
+          <Button variant="success">Learn</Button>
         </LinkContainer>
+        <LinkContainer to="/edit-deck-page" onClick={ () => setCurrentDeckId(deck.deckId) }>
+          <Button className="ml-3" variant="secondary">Edit</Button>
+        </LinkContainer>
+        <Button className="ml-3" variant="primary" onClick={() => exportDeck(deck.items, deck.deckTitle)}>
+          Export
+        </Button>
         <Button 
-          variant="danger" 
-          style={ { 'marginLeft': '20px' } }
+          className="ml-3"
+          variant="danger"
           onClick={ () => onDeleteDeck(deck.deckId) }
         >
           Delete
