@@ -1,4 +1,4 @@
-import { db } from '../db/firebase-config';
+import { db } from '../firebase/firebase-config';
 import { collection, getDocs, addDoc, query, where, updateDoc, getDoc, doc, deleteDoc } from 'firebase/firestore';
 import { DeckType, ItemType } from '../types/types';
 import { useState } from 'react';
@@ -7,6 +7,7 @@ const useFirestore = () => {
 
   const [loading, setLoading] = useState(false);
   const decksCollectionRef = collection(db, "decks");
+  const usersCollectionRef = collection(db, "users")
 
   const createDeck = async (deck: DeckType) => {
     await addDoc(decksCollectionRef, deck)
@@ -27,14 +28,29 @@ const useFirestore = () => {
     await updateDoc(deckRef, { items });
   };
 
+  const updateUserDeck = async (userId: string, item: string) => {
+    const userDocRef = doc(usersCollectionRef, userId)
+
+    await updateDoc(userDocRef, {
+      email: item
+    })
+  }
+
   const getDeckFromFirestore = async (deckId: string) => {
     setLoading(true);
     const deckRef = doc(decksCollectionRef, deckId);
     
     const querySnapshot = await getDoc(deckRef);
 
+    const data = { ...querySnapshot.data() }
+    
     setLoading(false);
-    return { ...querySnapshot.data(), deckId: querySnapshot.id }
+    return { 
+      deckTitle: data.deckTitle,
+      items: data.items,
+      userId: data.userId, 
+      deckId: querySnapshot.id 
+    }
   };
 
   const getUserDecksFromFirestore = async (userId: string) => {
@@ -43,7 +59,16 @@ const useFirestore = () => {
     const userDecks = query(decksCollectionRef, where("userId", "==", userId));
     const querySnapshot = await getDocs(userDecks);
 
-    decks = querySnapshot.docs.map((doc) => ({ ...doc.data(), deckId: doc.id }));
+    decks = querySnapshot.docs.map((doc) => {
+      const data = { ...doc.data() }
+
+      return ({ 
+        deckTitle: data.deckTitle,
+        items: data.items,
+        userId: data.userId,
+        deckId: doc.id 
+      })
+    });
     setLoading(false);
     return decks;
   }
@@ -54,6 +79,7 @@ const useFirestore = () => {
     updateDeck,
     getDeckFromFirestore,
     getUserDecksFromFirestore,
+    updateUserDeck,
     loading
   };
 
